Finish refactoring the `macros` module
