+++ src/macros.rs
@@ -4,6 +4,10 @@ The only public macro is `bitvec`; this module also provides convenience macros
 for code generation.
 !*/
 
+#[macro_use]
+#[doc(hidden)]
+pub mod internal;
+
 /** Construct a `&BitSlice` out of a literal array in source code, like `vec!`.
 
 `bits!` can be invoked in a number of ways. It takes the name of a `BitOrder`
@@ -38,299 +42,93 @@ bits![Lsb0; 0; 5];
 bits![1; 5];
 bits![bitvec::order::Local; 0, 1,];
 ```
-*/
+**/
 #[macro_export]
 macro_rules! bits {
+	//  Sequence syntax `[bit (, bit)*]` or `[(bit ,)*]`
+
+	//  Explicit order, literal `Word` store.
 	($order:ident, Word; $($val:expr),* $(,)?) => {
-		$crate::__bits_from_slice_word!($order; $($val),*)
+		$crate::__bits_from_words!($order; $($val),*)
 	};
+	/* Duplicate arms, differing in `$order:ident` and `$order:path`, force the
+	matcher to wrap a `:path`, which is `[:tt]`, as a single opaque `:tt` for
+	propagation through the macro call. Since the literal `$order` values will
+	match as `:ident`, not `:path`, this will only enter for orderings that the
+	rest of the macros would not be able to inspect and special-case *anyway*.
+	*/
 	($order:path, Word; $($val:expr),* $(,)?) => {
-		$crate::__bits_from_slice_word!($order; $($val),*)
+		$crate::__bits_from_words!($order; $($val),*)
 	};
-	($order:ident, $bits:ident; $($val:expr),* $(,)?) => {
-		$crate::__bits_from_slice!(
-			$order, $bits, $crate::__bits_count!($($val),*),
-			&$crate::__bits_store_array!($order, $bits; $($val),*)
-		)
+
+	//  Explicit order and store.
+	($order:ident, $store:ident; $($val:expr),* $(,)?) => {
+		&$crate::slice::BitSlice::<$order, $store>::from_slice(
+			&$crate::__bits_store_array!($order, $store; $($val),*)[..]
+		)[.. $crate::__count!($($val),*)]
 	};
-	($order:path, $bits:ident; $($val:expr),* $(,)?) => {
-		// Wrap multi-tt `order` into a single opaque tt.
-		$crate::__bits_from_slice!(
-			$order, $bits, $crate::__bits_count!($($val),*),
-			&$crate::__bits_store_array!($order, $bits; $($val),*)
-		)
+	($order:path, $store:ident; $($val:expr),* $(,)?) => {
+		&$crate::slice::BitSlice::<$order, $store>::from_slice(
+			&$crate::__bits_store_array!($order, $store; $($val),*)[..]
+		)[.. $crate::__count!($($val),*)]
 	};
+
+	//  Explicit order, default store.
 	($order:ident; $($val:expr),* $(,)?) => {
-		$crate::bits!($order, Word; $($val),*)
+		$crate::__bits_from_words!($order; $($val),*)
 	};
 	($order:path; $($val:expr),* $(,)?) => {
-		$crate::bits!($order, Word; $($val),*)
-	};
-	($($val:expr),* $(,)?) => {
-		$crate::bits!(Local, Word; $($val),*)
+		$crate::__bits_from_words!($order; $($val),*)
 	};
 
-	// NOTE: `len` must be `const`, as this is a non-allocating macro.
-	($order:ident, $bits:ident; $val:expr; $len:expr) => {
-		$crate::__bits_from_slice!($order, $bits, $len, &[
-			(0 as $bits).wrapping_sub(($val != 0) as $bits);
-			$crate::store::elts::<$bits>($len)
-		])
+	//  Default order and store.
+	($($val:expr),* $(,)?) => {
+		$crate::__bits_from_words!(Local; $($val),*)
 	};
-	($order:path, $bits:ident; $val:expr; $len:expr) => {
-		// Wrap multi-tt `order` into a single opaque tt.
-		$crate::__bits_from_slice!($order, $bits, $len, &[
-			(0 as $bits).wrapping_sub(($val != 0) as $bits);
-			$crate::store::elts::<$bits>($len)
-		])
+
+	//  Repetition syntax `[bit ; count]`
+	//  NOTE: `count` must be `const`, as this is a non-allocating macro.
+
+	//  Explicit order and store.
+	($order:ident, $store:ident; $val:expr; $len:expr) => {
+		&$crate::slice::BitSlice::<$order, $store>::from_slice(&[
+			(0 as $store).wrapping_sub(($val != 0) as $store)
+			; $crate::store::elts::<$store>($len)
+		][..])[.. $len]
 	};
+	($order:path, $store:ident; $val:expr; $len:expr) => {
+		&$crate::slice::BitSlice::<$order, $store>::from_slice(&[
+			(0 as $store).wrapping_sub(($val != 0) as $store)
+			; $crate::store::elts::<$store>($len)
+		][..])[.. $len]
+	};
+
+	//  Explicit order, default store.
 	($order:ident; $val:expr; $len:expr) => {
-		$crate::__bits_from_slice!($order, $crate::store::Word, $len, &[
+		&$crate::slice::BitSlice::<$order, $crate::store::Word>::from_slice(&[
 			(0 as $crate::store::Word).wrapping_sub(
 				($val != 0) as $crate::store::Word,
-			);
-			$crate::store::elts::<$crate::store::Word>($len)
-		])
+			); $crate::store::elts::<$crate::store::Word>($len)
+		][..])[.. $len]
 	};
 	($order:path; $val:expr; $len:expr) => {
-		$crate::__bits_from_slice!($order, $crate::store::Word, $len, &[
+		&$crate::slice::BitSlice::<$order, $crate::store::Word>::from_slice(&[
 			(0 as $crate::store::Word).wrapping_sub(
 				($val != 0) as $crate::store::Word,
-			);
-			$crate::store::elts::<$crate::store::Word>($len)
-		])
+			); $crate::store::elts::<$crate::store::Word>($len)
+		][..])[.. $len]
 	};
+
+	//  Default order and store.
 	($val:expr; $len:expr) => {
-		$crate::__bits_from_slice!(Local, $crate::store::Word, $len, &[
+		&$crate::slice::BitSlice::<
+			$crate::order::Local,
+			$crate::store::Word,
+		>::from_slice(&[
 			(0 as $crate::store::Word).wrapping_sub(
 				($val != 0) as $crate::store::Word,
-			);
-			$crate::store::elts::<$crate::store::Word>($len)
-		])
-	};
-}
-
-// Not public API
-#[doc(hidden)]
-#[macro_export]
-macro_rules! __bits_count {
-	(@ $val:expr) => {
-		1
-	};
-	($($val:expr),*) => {
-		0usize $(+ $crate::__bits_count!(@ $val))*
-	};
-}
-
-// Not public API
-#[doc(hidden)]
-#[macro_export]
-macro_rules! __bits_from_slice {
-	// Ensure literal `Msb0`, `Lsb0`, and `Local` always map to our internal
-	// definitions of the type, as we've been structurally matching them.
-	(Lsb0, $bits:ident, $len:expr, $slice:expr) => {
-		&$crate::slice::BitSlice::<
-			$crate::order::Lsb0,
-			$bits,
-		>::from_slice($slice)[.. $len]
-	};
-	(Msb0, $bits:ident, $len:expr, $slice:expr) => {
-		&$crate::slice::BitSlice::<
-			$crate::order::Msb0,
-			$bits,
-		>::from_slice($slice)[.. $len]
-	};
-	(Local, $bits:ident, $len:expr, $slice:expr) => {
-		&$crate::slice::BitSlice::<
-			$crate::order::Local,
-			$bits,
-		>::from_slice($slice)[.. $len]
-	};
-	(Local, $bits:path, $len:expr, $slice:expr) => {
-		&$crate::slice::BitSlice::<
-			$crate::order::Local,
-			$bits,
-		>::from_slice($slice)[.. $len]
-	};
-
-	// For other types, look it up in the caller's scope.
-	($order:tt, $bits:path, $len:expr, $slice:expr) => {
-		&$crate::slice::BitSlice::<$order, $bits>::from_slice($slice)[.. $len]
-	};
-}
-
-// Not public API
-#[doc(hidden)]
-#[macro_export]
-macro_rules! __elt_from_bits {
-	// Optimized Orders
-	(
-		Lsb0, $bits:ident;
-		$($a:tt, $b:tt, $c:tt, $d:tt, $e:tt, $f:tt, $g:tt, $h:tt),*
-	) => {
-		$bits::from_le_bytes([$($crate::private::u8_from_le_bits(
-			$a != 0, $b != 0, $c != 0, $d != 0,
-			$e != 0, $f != 0, $g != 0, $h != 0,
-		)),*])
-	};
-	(
-		Msb0, $bits:ident;
-		$($a:tt, $b:tt, $c:tt, $d:tt, $e:tt, $f:tt, $g:tt, $h:tt),*
-	) => {
-		$bits::from_be_bytes([$($crate::private::u8_from_be_bits(
-			$a != 0, $b != 0, $c != 0, $d != 0,
-			$e != 0, $f != 0, $g != 0, $h != 0,
-		)),*])
-	};
-	(
-		Local, $bits:ident;
-		$($a:tt, $b:tt, $c:tt, $d:tt, $e:tt, $f:tt, $g:tt, $h:tt),*
-	) => {
-		$bits::from_ne_bytes([$($crate::private::u8_from_local_bits(
-			$a != 0, $b != 0, $c != 0, $d != 0,
-			$e != 0, $f != 0, $g != 0, $h != 0,
-		)),*])
-	};
-
-	// Arbitrary Order
-	($order:tt, $bits:ident; $($a:tt),*) => {
-		{
-			let mut value: $bits = 0;
-			let _idx = 0u8;
-			$(
-				$crate::store::BitStore::set::<$order>(
-					&mut value,
-					unsafe { $crate::indices::BitIdx::new_unchecked(_idx) },
-					$a != 0,
-				);
-				let _idx = _idx + 1;
-			)*
-			value
-		}
-	};
-}
-
-// Not public API
-#[doc(hidden)]
-#[macro_export]
-macro_rules! __bits_store_array {
-	// Entry point.
-	($order:tt, $bits:ident; $($val:expr),*) => {
-		$crate::__bits_store_array!(
-			$order, $bits, []; $($val,)*
-			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 16
-			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 32
-			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 48
-			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0  // 64
-		);
-	};
-
-	// NOTE: This has to be first. Abuses quirk of `macro_rules!` where :expr
-	// captures are considered a single tt after being matched. Even if the
-	// :expr matcher was a literal `0`, after being wrapped by the :expr NT, it
-	// no longer is considered to match a literal `0`, so this pattern will only
-	// match the extra `0`s added at the end.
-	($order:tt, $bits:ident, [$( ($($elt:tt),*) )*]; $(0),*) => {
-		[$(
-			$crate::__elt_from_bits!($order, $bits; $($elt),*)
-		),*]
-	};
-
-	// Matchers for each size of word. The end of the word may be padded out
-	// with `0`s.
-	(
-		$order:tt, u8, [$($w:tt)*];
-		$a0:tt, $b0:tt, $c0:tt, $d0:tt, $e0:tt, $f0:tt, $g0:tt, $h0:tt
-		$(, $($t:tt)*)?
-	) => {
-		$crate::__bits_store_array!(
-			$order, u8, [$($w)* (
-				$a0, $b0, $c0, $d0, $e0, $f0, $g0, $h0
-			)];
-			$($($t)*)?
-		);
-	};
-	(
-		$order:tt, u16, [$($w:tt)*];
-		$a0:tt, $b0:tt, $c0:tt, $d0:tt, $e0:tt, $f0:tt, $g0:tt, $h0:tt,
-		$a1:tt, $b1:tt, $c1:tt, $d1:tt, $e1:tt, $f1:tt, $g1:tt, $h1:tt
-		$(, $($t:tt)*)?
-	) => {
-		$crate::__bits_store_array!(
-			$order, u16, [$($w)* (
-				$a0, $b0, $c0, $d0, $e0, $f0, $g0, $h0,
-				$a1, $b1, $c1, $d1, $e1, $f1, $g1, $h1
-			)];
-			$($($t)*)?
-		);
-	};
-	(
-		$order:tt, u32, [$($w:tt)*];
-		$a0:tt, $b0:tt, $c0:tt, $d0:tt, $e0:tt, $f0:tt, $g0:tt, $h0:tt,
-		$a1:tt, $b1:tt, $c1:tt, $d1:tt, $e1:tt, $f1:tt, $g1:tt, $h1:tt,
-		$a2:tt, $b2:tt, $c2:tt, $d2:tt, $e2:tt, $f2:tt, $g2:tt, $h2:tt,
-		$a3:tt, $b3:tt, $c3:tt, $d3:tt, $e3:tt, $f3:tt, $g3:tt, $h3:tt
-		$(, $($t:tt)*)?
-	) => {
-		$crate::__bits_store_array!(
-			$order, u32, [$($w)* (
-				$a0, $b0, $c0, $d0, $e0, $f0, $g0, $h0,
-				$a1, $b1, $c1, $d1, $e1, $f1, $g1, $h1,
-				$a2, $b2, $c2, $d2, $e2, $f2, $g2, $h2,
-				$a3, $b3, $c3, $d3, $e3, $f3, $g3, $h3
-			)];
-			$($($t)*)?
-		);
-	};
-	(
-		$order:tt, u64, [$($w:tt)*];
-		$a0:tt, $b0:tt, $c0:tt, $d0:tt, $e0:tt, $f0:tt, $g0:tt, $h0:tt,
-		$a1:tt, $b1:tt, $c1:tt, $d1:tt, $e1:tt, $f1:tt, $g1:tt, $h1:tt,
-		$a2:tt, $b2:tt, $c2:tt, $d2:tt, $e2:tt, $f2:tt, $g2:tt, $h2:tt,
-		$a3:tt, $b3:tt, $c3:tt, $d3:tt, $e3:tt, $f3:tt, $g3:tt, $h3:tt,
-		$a4:tt, $b4:tt, $c4:tt, $d4:tt, $e4:tt, $f4:tt, $g4:tt, $h4:tt,
-		$a5:tt, $b5:tt, $c5:tt, $d5:tt, $e5:tt, $f5:tt, $g5:tt, $h5:tt,
-		$a6:tt, $b6:tt, $c6:tt, $d6:tt, $e6:tt, $f6:tt, $g6:tt, $h6:tt,
-		$a7:tt, $b7:tt, $c7:tt, $d7:tt, $e7:tt, $f7:tt, $g7:tt, $h7:tt
-		$(, $($t:tt)*)?
-	) => {
-		$crate::__bits_store_array!(
-			$order, u64, [$($w)* (
-				$a0, $b0, $c0, $d0, $e0, $f0, $g0, $h0,
-				$a1, $b1, $c1, $d1, $e1, $f1, $g1, $h1,
-				$a2, $b2, $c2, $d2, $e2, $f2, $g2, $h2,
-				$a3, $b3, $c3, $d3, $e3, $f3, $g3, $h3,
-				$a4, $b4, $c4, $d4, $e4, $f4, $g4, $h4,
-				$a5, $b5, $c5, $d5, $e5, $f5, $g5, $h5,
-				$a6, $b6, $c6, $d6, $e6, $f6, $g6, $h6,
-				$a7, $b7, $c7, $d7, $e7, $f7, $g7, $h7
-			)];
-			$($($t)*)?
-		);
-	};
-}
-
-#[doc(hidden)]
-#[macro_export]
-#[cfg(target_pointer_width = "32")]
-macro_rules! __bits_from_slice_word {
-	($order:tt; $($val:expr),*) => {
-		$crate::__bits_from_slice!(
-			$order, u32, $crate::__bits_count!($($val),*),
-			&$crate::__bits_store_array!($order, u32; $($val),*)
-		)
-	};
-}
-
-#[doc(hidden)]
-#[macro_export]
-#[cfg(target_pointer_width = "64")]
-macro_rules! __bits_from_slice_word {
-	($order:tt; $($val:expr),*) => {
-		$crate::__bits_from_slice!(
-			$order, u64, $crate::__bits_count!($($val),*),
-			&$crate::__bits_store_array!($order, u64; $($val),*)
-		)
+			); $crate::store::elts::<$crate::store::Word>($len)
+		][..])[.. $len]
 	};
 }
 
@@ -384,47 +182,46 @@ bitvec![1; 5];
 #[cfg(feature = "alloc")]
 #[macro_export]
 macro_rules! bitvec {
-	($order:ident, $bits:ident; $($val:expr),* $(,)?) => {
-		{
-			let mut bitvec = $crate::__bitvec_from_slice!(
-				$order,
-				$bits,
-				$crate::__bits_store_array!($order, $bits; $($val),*)
-			);
-			bitvec.truncate($crate::__bits_count!($($val),*));
-			bitvec
-		}
+	/* TODO(myrrlyn): Make this arm viable.
+	($($arg:tt)) => {
+		($crate::bits!($($arg))).to_owned()
 	};
-	($order:path, $bits:ident; $($val:expr),* $(,)?) => {
-		// Wrap multi-tt `order` into a single opaque tt.
-		{
-			let mut bitvec = $crate::__bitvec_from_slice!(
-				$order,
-				$bits,
-				$crate::__bits_store_array!($order, $bits; $($val),*)
-			);
-			bitvec.truncate($crate::__bits_count!($($val),*));
-			bitvec
-		}
+	*/
+
+	//  Sequence syntax
+
+	($order:ident, $store:ident; $($val:expr),* $(,)?) => {
+		$crate::vec::BitVec::<$order, $store>::from_bitslice(
+			$crate::bits!($order, $store; $($val),*),
+		)
 	};
+	($order:path, $store:ident; $($val:expr),* $(,)?) => {
+		$crate::vec::BitVec::<$order, $store>::from_bitslice(
+			$crate::bits!($order, $store; $($val),*),
+		)
+	};
+
 	($order:ident; $($val:expr),* $(,)?) => {
 		$crate::vec::BitVec::<
 			$order,
 			$crate::store::Word,
-		>::from_bitslice($crate::__bits_from_slice_word!($order; $($val),*))
+		>::from_bitslice($crate::__bits_from_words!($order; $($val),*))
 	};
 	($order:path; $($val:expr),* $(,)?) => {
 		$crate::vec::BitVec::<
 			$order,
 			$crate::store::Word,
-		>::from_bitslice($crate::__bits_from_slice_word!($order; $($val),*))
+		>::from_bitslice($crate::__bits_from_words!($order; $($val),*))
 	};
+
 	($($val:expr),* $(,)?) => {
 		$crate::vec::BitVec::<
 			$crate::order::Local,
 			$crate::store::Word,
-		>::from_bitslice($crate::__bits_from_slice_word!(Local; $($val),*))
+		>::from_bitslice($crate::__bits_from_words!(Local; $($val),*))
 	};
+
+	//  Repetition syntax
 
 	($order:ident, $bits:ident; $val:expr; $len:expr) => {
 		$crate::vec::BitVec::<$order, $bits>::repeat($val != 0, $len)
@@ -432,6 +229,7 @@ macro_rules! bitvec {
 	($order:path, $bits:ident; $val:expr; $len:expr) => {
 		$crate::vec::BitVec::<$order, $bits>::repeat($val != 0, $len)
 	};
+
 	($order:ident; $val:expr; $len:expr) => {
 		$crate::vec::BitVec::<
 			$order,
@@ -444,43 +242,12 @@ macro_rules! bitvec {
 			$crate::store::Word,
 		>::repeat($val != 0, $len)
 	};
+
 	($val:expr; $len:expr) => {
 		$crate::vec::BitVec::<
 			$crate::order::Local,
 			$crate::store::Word,
 		>::repeat($val != 0, $len)
-	};
-}
-
-// Not public API
-#[cfg(feature = "alloc")]
-#[doc(hidden)]
-#[macro_export]
-macro_rules! __bitvec_from_slice {
-	// Ensure literal `Msb0`, `Lsb0`, and `Local` always map to our internal
-	// definitions of the type, as we've been structurally matching them.
-	(Lsb0, $bits:ident, $slice:expr) => {
-		$crate::vec::BitVec::<
-			$crate::order::Lsb0,
-			$bits,
-		>::from_slice(&$slice[..])
-	};
-	(Msb0, $bits:ident, $slice:expr) => {
-		$crate::vec::BitVec::<
-			$crate::order::Msb0,
-			$bits,
-		>::from_slice(&$slice[..])
-	};
-	(Local, $bits:ident, $slice:expr) => {
-		$crate::vec::BitVec::<
-			$crate::order::Local,
-			$bits,
-		>::from_slice(&$slice[..])
-	};
-
-	// For other types, look it up in the caller's scope.
-	($order:tt, $bits:tt, $slice:expr) => {
-		$crate::vec::BitVec::<$order, $bits>::from_slice(&$slice[..])
 	};
 }
 
@@ -495,90 +262,9 @@ freeze it.
 #[cfg(feature = "alloc")]
 #[macro_export]
 macro_rules! bitbox {
-	($($t:tt)*) => {
-		$crate::bitvec![$($t)*].into_boxed_bitslice()
+	($($arg:tt)*) => {
+		$crate::bitvec![$($arg)*].into_boxed_bitslice()
 	};
-}
-
-#[doc(hidden)]
-macro_rules! __bitslice_shift {
-	( $( $t:ty ),+ ) => { $(
-		#[doc(hidden)]
-		impl<C, T >core::ops::ShlAssign<$t>
-		for $crate::prelude::BitSlice<C,T>
-		where C: $crate::order::BitOrder, T: $crate::store::BitStore {
-			fn shl_assign(&mut self, shamt: $t) {
-				core::ops::ShlAssign::<usize>::shl_assign(
-					self,
-					shamt as usize,
-				)
-			}
-		}
-
-		#[doc(hidden)]
-		impl<C, T> core::ops::ShrAssign<$t>
-		for $crate::prelude::BitSlice<C,T>
-		where C: $crate::order::BitOrder, T: $crate::store::BitStore {
-			fn shr_assign(&mut self,shamt: $t){
-				core::ops::ShrAssign::<usize>::shr_assign(
-					self,
-					shamt as usize,
-				)
-			}
-		}
-	)+ };
-}
-
-#[cfg(feature = "alloc")]
-#[doc(hidden)]
-macro_rules! __bitvec_shift {
-	( $( $t:ty ),+ ) => { $(
-		#[doc(hidden)]
-		impl<C, T> core::ops::Shl<$t>
-		for $crate::vec::BitVec<C, T>
-		where C: $crate::order::BitOrder, T: $crate::store::BitStore {
-			type Output = <Self as core::ops::Shl<usize>>::Output;
-
-			fn shl(self, shamt: $t) -> Self::Output {
-				core::ops::Shl::<usize>::shl(self, shamt as usize)
-			}
-		}
-
-		#[doc(hidden)]
-		impl<C, T> core::ops::ShlAssign<$t>
-		for $crate::vec::BitVec<C, T>
-		where C: $crate::order::BitOrder, T: $crate::store::BitStore {
-			fn shl_assign(&mut self, shamt: $t) {
-				core::ops::ShlAssign::<usize>::shl_assign(
-					self,
-					shamt as usize,
-				)
-			}
-		}
-
-		#[doc(hidden)]
-		impl<C, T> core::ops::Shr<$t>
-		for $crate::vec::BitVec<C, T>
-		where C: $crate::order::BitOrder, T: $crate::store::BitStore {
-			type Output = <Self as core::ops::Shr<usize>>::Output;
-
-			fn shr(self, shamt: $t) -> Self::Output {
-				core::ops::Shr::<usize>::shr(self, shamt as usize)
-			}
-		}
-
-		#[doc(hidden)]
-		impl<C, T> core::ops::ShrAssign<$t>
-		for $crate::vec::BitVec<C, T>
-		where C: $crate::order::BitOrder, T: $crate::store::BitStore {
-			fn shr_assign(&mut self, shamt: $t) {
-				core::ops::ShrAssign::<usize>::shr_assign(
-					self,
-					shamt as usize,
-				)
-			}
-		}
-	)+ };
 }
 
 #[cfg(test)]
